+++ script.video.smartishplaylist/default.py
@@ -1,48 +1,90 @@
 import sys
 import json
-
+import threading
+import xbmc
+from queue import Queue
 import xbmcaddon
 import xbmcgui
 
 from resources.lib.logger import write_log
-from resources.lib.selections import select_media, display_selections
-from resources.lib.playlist_functions import build_playlist, gather_media_info, start_playlist, quit_kodi_after
+from resources.lib.selections import select_media, configure_shows
+from resources.lib.playlist_functions import gather_media_info, quit_kodi_after, playlist_builder, video_playlist_start
 
 all_args = sys.argv
 
 write_log(f"all args: {all_args}")
 
 
-def build():
-    addon = xbmcaddon.Addon()
-
-    autoplay = addon.getSettingBool("auto_play")
-    shuffle = addon.getSettingBool("shuffle")
-    auto_quit = addon.getSettingBool("auto_quit")
-
-    write_log(f"Building, Autoplay: {autoplay} Shuffle: {shuffle}")
-
-    progress = xbmcgui.DialogProgress()
-    progress.create("Building Playlist", "Initializing...")
-    items = gather_media_info()
-    playlist_progress = build_playlist(media_info=items, progress_dialog=progress)
-    progress.close()
+def rpc_worker(progress_queue:Queue, cancel_event:threading.Event) -> None:
+    write_log("Begin RPC worker")
+    monitor = xbmc.Monitor()
+    items = gather_media_info(monitor=monitor)
+    playlist_progress = playlist_builder(media_info=items, monitor=monitor, progress_queue=progress_queue, cancel_event=cancel_event)
 
     if playlist_progress:
-        xbmcgui.Dialog().notification("Playlist Ready", "Build complete", xbmcgui.NOTIFICATION_INFO, 3000)
         write_log("Playlist build complete")
 
-        if autoplay:
-            start_playlist(playlist_id=1, shuffle=shuffle)
-            write_log("Playback started")
 
-            if auto_quit:
-                auto_quit_minutes = json.loads(addon.getSetting("auto_quit_minutes"))
-                quit_kodi_after(auto_quit_minutes)
+def run() -> None:
+        addon = xbmcaddon.Addon()
+        progress = xbmcgui.DialogProgress()
 
-    else:
-        xbmcgui.Dialog().notification("Stopped", "Playlist build incomplete", xbmcgui.NOTIFICATION_ERROR, 3000)
-        write_log("Playlist build cancelled")
+        autoplay = addon.getSettingBool("auto_play")
+        shuffle = addon.getSettingBool("shuffle")
+        auto_quit = addon.getSettingBool("auto_quit")
+
+        write_log(f"Building, Autoplay: {autoplay} Shuffle: {shuffle}")
+
+        progress.create("Building Playlist", "Initializing...")
+
+        progress_queue = Queue()
+        cancel_event = threading.Event()
+
+        background_worker = threading.Thread(
+            target=rpc_worker,
+            args=(progress_queue, cancel_event),
+            daemon=True
+        )
+
+        background_worker.start()
+        cancelled = False
+
+        while True:
+            if progress.iscanceled():
+                cancel_event.set()
+                cancelled = True
+                break
+
+            try:
+                message = progress_queue.get_nowait()
+
+                if message:
+                    if message[0] == "done":
+                        break
+
+                    percent, text = message
+                    progress.update(percent, text)
+
+            except Exception as e:
+                write_log(f"Error: {e}")
+
+            xbmc.sleep(50)
+
+        progress.close()
+
+        if cancelled:
+            xbmcgui.Dialog().notification("Stopped", "Playlist build incomplete", xbmcgui.NOTIFICATION_ERROR, 3000)
+            write_log("Playlist build cancelled")
+
+        else:
+            xbmcgui.Dialog().notification("Playlist Ready", "Build complete", xbmcgui.NOTIFICATION_INFO, 3000)
+            if autoplay:
+                video_playlist_start(shuffle=shuffle)
+                write_log("Playback started")
+
+                if auto_quit:
+                    auto_quit_minutes = json.loads(addon.getSetting("auto_quit_minutes"))
+                    quit_kodi_after(auto_quit_minutes)
 
 
 def main():
@@ -55,11 +97,11 @@ def main():
         if action == "select_movies":
             select_media(media_type="movie")
 
-        if action == "selection_info":
-            display_selections()
+        if action == "configure_shows":
+            configure_shows()
 
     else:
-        build()
+        run()
 
 
 if __name__ == "__main__":