+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+#### 0.3.0
+
+---
+
+- GUI modification to allow for individual TV Show configuration.
+- Moved RPC calls to background thread
+- Removed "Current Selections" from menu
+- Added icon.png and screenshots
+- Selection info now stored in local JSON (special://profile/addon_data/script.video.smartishplaylist/config.json)
+
 #### 0.2.1
 
 ---